moving search into own package and renaming funcs throughout
